requirements.txt
Django >4.0
requests
mysqlclient
aiohttp
djangorestframework
aioboto3
redis 
httpx
